+++ ofrak_core/src/ofrak/resource.py
@@ -529,9 +529,10 @@ class Resource:
     def get_attributes(self, attributes_type: Type[RA]) -> RA:
         """
         If this resource has attributes matching the given type, return the value of those
-        attributes. Otherwise returns `None`.
+        attributes. Otherwise raise `NotFoundError`.
         :param attributes_type:
         :return:
+        :raises NotFoundError: if the resource does not have the requested attributes
         """
         if attributes_type not in self.attributes:
             raise NotFoundError()